Correctly specify a JSON request body payload as a single schema
Currently, a JSON object payload is treated as a collection of separate parameters and specified as a list of schema in the generated Swagger spec. This commit changes this behaviour to comply with https://swagger.io/docs/specification/2-0/describing-request-body/ where object payloads are specified using a single schema encapsulation all the properties.